requirements.txt
alembic==1.7.4
click==8.0.3
flake8==4.0.1
Flask==2.0.2
Flask-HTTPAuth==4.4.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.2
Mako==1.1.5
MarkupSafe==2.0.1
mccabe==0.6.1
psycopg2-binary==2.9.1
pycodestyle==2.8.0
pyflakes==2.4.0
python-dotenv==0.19.1
SQLAlchemy==1.4.25
Werkzeug==2.0.2
